backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.143.0
python-dotenv==1.0.1
requests==2.32.3
